Update __init__.py
Added an optional "merge distance" setting to the Generate Collision Mesh operator. At the default setting, there is no change to how the operator worked before. But increasing the value of this setting will cause the resulting collision mesh to be much lower-poly and less complex, so it's a great way to optimize the mesh further.