Update __init__.py
Added a new QC override system that allows the user to customize the resulting generated QC files, with custom QC commands (such as $scale, $surfaceprop, etc.) The overrides are added to collision objects in Blender via Custom Properties. Any string-based Custom Property that has a name starting with "$" will be treated as a QC override.